+++ src/extension/module.js
@@ -1432,6 +1432,14 @@
       }
     };
 
+    const isChangedUrl = () => {
+      const $test = document.getElementById('sidekick_test_location');
+      if ($test) {
+        return $test.value !== sk.location.href;
+      }
+      return window.location.href !== sk.location.href;
+    };
+
     const updateBulkInfo = () => {
       if (!sk.isAdmin()) {
         return;
@@ -1449,7 +1457,7 @@
       }
       // show/hide bulk buttons
       const filesSelected = sel.length > 0;
-      if (window.location.href !== sk.location.href) {
+      if (isChangedUrl()) {
         // refresh location
         sk.location = getLocation();
       }
@@ -1480,14 +1488,6 @@
       return i18n(sk, i18nKey)
         .replace('$1', num)
         .replace('$2', total);
-    };
-
-    const isChangedUrl = () => {
-      const $test = document.getElementById('sidekick_test_location');
-      if ($test) {
-        return $test.value !== sk.location.href;
-      }
-      return window.location.href !== sk.location.href;
     };
 
     const doBulkOperation = async (operation, method, concurrency, host) => {
@@ -1591,15 +1591,10 @@
             sk.showModal(confirmText);
           } else if (window.confirm(confirmText)) {
             sk.showWait();
-            if (isChangedUrl()) {
-              // url changed, refetch status
-              sk.addEventListener('statusfetched', () => {
-                doBulkOperation('preview', 'update', 2, sk.config.innerHost);
-              }, { once: true });
-              sk.fetchStatus(true);
-            } else {
+            sk.addEventListener('statusfetched', () => {
               doBulkOperation('preview', 'update', 2, sk.config.innerHost);
-            }
+            }, { once: true });
+            sk.fetchStatus(true);
           }
         },
         isEnabled: (s) => s.isAuthorized('preview', 'write') && s.status.webPath,
@@ -1618,15 +1613,10 @@
             sk.showModal(confirmText);
           } else if (window.confirm(confirmText)) {
             sk.showWait();
-            if (isChangedUrl()) {
-              // url changed, refetch status
-              sk.addEventListener('statusfetched', () => {
-                doBulkOperation('publish', 'publish', 40, sk.config.host || sk.config.outerHost);
-              }, { once: true });
-              sk.fetchStatus(true);
-            } else {
+            sk.addEventListener('statusfetched', () => {
               doBulkOperation('publish', 'publish', 40, sk.config.host || sk.config.outerHost);
-            }
+            }, { once: true });
+            sk.fetchStatus(true);
           }
         },
         isEnabled: (s) => s.isAuthorized('live', 'write') && s.status.webPath,
@@ -1673,15 +1663,10 @@
               sk.showModal(emptyText);
             } else {
               sk.showWait();
-              if (isChangedUrl()) {
-                // url changed, refetch status
-                sk.addEventListener('statusfetched', () => {
-                  doBulkCopyUrls(hostProperty);
-                }, { once: true });
-                sk.fetchStatus(true);
-              } else {
+              sk.addEventListener('statusfetched', () => {
                 doBulkCopyUrls(hostProperty);
-              }
+              }, { once: true });
+              sk.fetchStatus(true);
             }
           },
         },
@@ -3543,13 +3528,11 @@
         // bust client cache for live and production
         if (config.outerHost) {
           // reuse purgeURL to ensure page relative paths (e.g. when publishing dependencies)
-          purgeURL.hostname = config.outerHost;
-          await fetch(purgeURL.href, { cache: 'reload', mode: 'no-cors' });
+          await fetch(`https://${config.outerHost}${purgeURL.pathname}`, { cache: 'reload', mode: 'no-cors' });
         }
         if (config.host) {
           // reuse purgeURL to ensure page relative paths (e.g. when publishing dependencies)
-          purgeURL.hostname = config.host;
-          await fetch(purgeURL.href, { cache: 'reload', mode: 'no-cors' });
+          await fetch(`https://${config.host}${purgeURL.pathname}`, { cache: 'reload', mode: 'no-cors' });
         }
         fireEvent(this, 'published', path);
       } catch (e) {